Refatora Gestão da Base em módulos
